Исправил и доработал
