+++ templates/main.php
@@ -1,27 +1,13 @@
-<?php
-
-function lot_price($x)
-{
-    $x = ceil($x);
-    if ($x >= 1000) {
-        $x = number_format($x, 0, '.', ' ');
-    }
-
-    return $x . ' ' . '₽';
-}
-?>
-
 <main class="container">
     <section class="promo">
         <h2 class="promo__title">Нужен стафф для катки?</h2>
         <p class="promo__text">На нашем интернет-аукционе ты найдёшь самое эксклюзивное сноубордическое и горнолыжное снаряжение.</p>
         <ul class="promo__list">
-            <?php $categories; ?>
-            <!--заполните этот список из массива категорий-->
+
             <li class="promo__item promo__item--boards">
                 <?php
                 foreach ($categories as $val) : ?>
-                    <a class="promo__link" href="pages/all-lots.html"><?=htmlspecialchars($val);?></a>
+                    <a class="promo__link" href="pages/all-lots.html"><?= htmlspecialchars($val); ?></a>
                 <?php endforeach; ?>
             </li>
         </ul>
@@ -31,20 +17,19 @@ function lot_price($x)
             <h2>Открытые лоты</h2>
         </div>
         <ul class="lots__list">
-            <?php $staff; ?>
-            <!--заполните этот список из массива с товарами-->
             <li class="lots__item lot">
                 <?php foreach ($staff as $key => $val) : ?>
                     <div class="lot__image">
-                        <img src="<?= $val['image']; ?>" width="350" height="260" alt="">
+                        <img src="<?= htmlspecialchars($val['image']); ?>" width="350" height="260" alt="">
                     </div>
                     <div class="lot__info">
-                        <span class="lot__category"><?= $val['category']; ?></span>
-                        <h3 class="lot__title"><a class="text-link" href="pages/lot.html"><?= $val['name']; ?></a></h3>
+                        <span class="lot__category"><?= htmlspecialchars($val['category']); ?></span>
+                        <h3 class="lot__title"><a class="text-link" href="pages/lot.html"><?= htmlspecialchars($val['name']); ?></a></h3>
                         <div class="lot__state">
                             <div class="lot__rate">
                                 <span class="lot__amount">Стартовая цена</span>
-                                <span class="lot__cost"><?= lot_price($val['price']); ?></span>
+                                <?php require_once('index.php') ?>
+                                <span class="lot__cost"><?= htmlspecialchars(lot_price($val['price'])); ?></span>
                             </div>
                             <div class="lot__timer timer">
                                 12:23
